+++ covenant/covenant_test.go
@@ -2,17 +2,21 @@ package covenant_test
 
 import (
 	"encoding/hex"
+	"fmt"
 	"math/rand"
 	"testing"
-
-	"github.com/btcsuite/btcd/btcutil"
+	"time"
 
 	"github.com/babylonlabs-io/babylon/btcstaking"
 	asig "github.com/babylonlabs-io/babylon/crypto/schnorr-adaptor-signature"
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	bbntypes "github.com/babylonlabs-io/babylon/types"
+	"github.com/btcsuite/btcd/btcec/v2"
+	"github.com/btcsuite/btcd/btcec/v2/schnorr"
+	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg"
 	"github.com/btcsuite/btcd/wire"
+	"github.com/decred/dcrd/dcrec/secp256k1/v4"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 
@@ -203,4 +207,178 @@ func FuzzAddCovenantSig(f *testing.F) {
 		require.NoError(t, err)
 		require.Equal(t, expectedTxHash, res.TxHash)
 	})
+}
+
+func TestDeduplicationWithOddKey(t *testing.T) {
+	// 1. Public key with odd y coordinate
+	oddKey := "0379a71ffd71c503ef2e2f91bccfc8fcda7946f4653cef0d9f3dde20795ef3b9f0"
+	oddKeyBytes, err := hex.DecodeString(oddKey)
+	require.NoError(t, err)
+	oddKeyPub, err := btcec.ParsePubKey(oddKeyBytes)
+	require.NoError(t, err)
+
+	// 2. Serialize in BTC schnorr format
+	serializedOddKey := schnorr.SerializePubKey(oddKeyPub)
+	pubKeyFromSchnorr, err := schnorr.ParsePubKey(serializedOddKey)
+	require.NoError(t, err)
+
+	randomKey, err := btcec.NewPrivateKey()
+	require.NoError(t, err)
+	pubKey := randomKey.PubKey()
+
+	paramVersion := uint32(2)
+	delegations := []*types.Delegation{
+		&types.Delegation{
+			CovenantSigs: []*types.CovenantAdaptorSigInfo{
+				&types.CovenantAdaptorSigInfo{
+					// 3. Delegation is already signed by the public key with odd y coordinate
+					Pk: pubKeyFromSchnorr,
+				},
+			},
+			ParamsVersion: paramVersion,
+		},
+		&types.Delegation{
+			CovenantSigs: []*types.CovenantAdaptorSigInfo{
+				&types.CovenantAdaptorSigInfo{
+					Pk: pubKey,
+				},
+			},
+			ParamsVersion: paramVersion,
+		},
+	}
+
+	paramsGet := NewMockParam(map[uint32]*types.StakingParams{
+		paramVersion: &types.StakingParams{
+			CovenantPks: []*secp256k1.PublicKey{oddKeyPub, pubKeyFromSchnorr},
+		},
+	})
+
+	// 4. After removing the already signed delegation, the list should have only one element
+	sanitized, err := covenant.SanitizeDelegations(oddKeyPub, paramsGet, delegations)
+	require.Equal(t, 1, len(sanitized))
+	require.NoError(t, err)
+}
+
+func TestIsKeyInCommittee(t *testing.T) {
+	r := rand.New(rand.NewSource(time.Now().Unix()))
+
+	// create a Covenant key pair in the keyring
+	covenantConfig := covcfg.DefaultConfig()
+	covenantConfig.BabylonConfig.KeyDirectory = t.TempDir()
+
+	covKeyPair, err := keyring.CreateCovenantKey(
+		covenantConfig.BabylonConfig.KeyDirectory,
+		covenantConfig.BabylonConfig.ChainID,
+		covenantConfig.BabylonConfig.Key,
+		covenantConfig.BabylonConfig.KeyringBackend,
+		passphrase,
+		hdPath,
+	)
+	require.NoError(t, err)
+	covenantSerializedPk := schnorr.SerializePubKey(covKeyPair.PublicKey)
+
+	// create params and version
+	pVersionWithoutCovenant := uint32(datagen.RandomInRange(r, 1, 10))
+	pVersionWithCovenant := pVersionWithoutCovenant + 1
+
+	paramsWithoutCovenant := testutil.GenRandomParams(r, t)
+	paramsWithCovenant := testutil.GenRandomParams(r, t)
+	paramsWithCovenant.CovenantPks = append(paramsWithCovenant.CovenantPks, covKeyPair.PublicKey)
+
+	// creates delegations to check
+	delNoCovenant := &types.Delegation{
+		ParamsVersion: pVersionWithoutCovenant,
+	}
+	delWithCovenant := &types.Delegation{
+		ParamsVersion: pVersionWithCovenant,
+	}
+
+	// simple mock with the parameter versions
+	paramsGet := NewMockParam(map[uint32]*types.StakingParams{
+		pVersionWithoutCovenant: paramsWithoutCovenant,
+		pVersionWithCovenant:    paramsWithCovenant,
+	})
+
+	// checks the case where the covenant is NOT in the committee
+	actual, err := covenant.IsKeyInCommittee(paramsGet, covenantSerializedPk, delNoCovenant)
+	require.False(t, actual)
+	require.NoError(t, err)
+	emptyDels, err := covenant.SanitizeDelegations(covKeyPair.PublicKey, paramsGet, []*types.Delegation{delNoCovenant, delNoCovenant})
+	require.NoError(t, err)
+	require.Len(t, emptyDels, 0)
+
+	// checks the case where the covenant is in the committee
+	actual, err = covenant.IsKeyInCommittee(paramsGet, covenantSerializedPk, delWithCovenant)
+	require.True(t, actual)
+	require.NoError(t, err)
+	dels, err := covenant.SanitizeDelegations(covKeyPair.PublicKey, paramsGet, []*types.Delegation{delWithCovenant, delNoCovenant})
+	require.NoError(t, err)
+	require.Len(t, dels, 1)
+	dels, err = covenant.SanitizeDelegations(covKeyPair.PublicKey, paramsGet, []*types.Delegation{delWithCovenant})
+	require.NoError(t, err)
+	require.Len(t, dels, 1)
+
+	amtSatFirst := btcutil.Amount(100)
+	amtSatSecond := btcutil.Amount(150)
+	amtSatThird := btcutil.Amount(200)
+	lastUnsanitizedDels := []*types.Delegation{
+		&types.Delegation{
+			ParamsVersion: pVersionWithCovenant,
+			TotalSat:      amtSatFirst,
+		},
+		delNoCovenant,
+		&types.Delegation{
+			ParamsVersion: pVersionWithCovenant,
+			TotalSat:      amtSatSecond,
+		},
+		delNoCovenant,
+		&types.Delegation{
+			ParamsVersion: pVersionWithCovenant,
+			TotalSat:      amtSatThird,
+		},
+	}
+
+	sanitizedDels, err := covenant.SanitizeDelegations(covKeyPair.PublicKey, paramsGet, lastUnsanitizedDels)
+	require.NoError(t, err)
+	require.Len(t, sanitizedDels, 3)
+	require.Equal(t, amtSatFirst, sanitizedDels[0].TotalSat)
+	require.Equal(t, amtSatSecond, sanitizedDels[1].TotalSat)
+	require.Equal(t, amtSatThird, sanitizedDels[2].TotalSat)
+
+	errParamGet := fmt.Errorf("dumbErr")
+	sanitizedDels, err = covenant.SanitizeDelegations(covKeyPair.PublicKey, NewMockParamError(errParamGet), lastUnsanitizedDels)
+	require.Nil(t, sanitizedDels)
+
+	errKeyIsInCommittee := fmt.Errorf("unable to get the param version: %d, reason: %s", pVersionWithCovenant, errParamGet.Error())
+	expErr := fmt.Errorf("unable to verify if covenant key is in committee: %s", errKeyIsInCommittee.Error())
+	require.EqualError(t, err, expErr.Error())
+}
+
+type MockParamGetter struct {
+	paramsByVersion map[uint32]*types.StakingParams
+}
+
+func NewMockParam(p map[uint32]*types.StakingParams) *MockParamGetter {
+	return &MockParamGetter{
+		paramsByVersion: p,
+	}
+}
+
+func (m *MockParamGetter) Get(version uint32) (*types.StakingParams, error) {
+	p := m.paramsByVersion[version]
+	return p, nil
+}
+
+type MockParamError struct {
+	err error
+}
+
+func NewMockParamError(err error) *MockParamError {
+	return &MockParamError{
+		err: err,
+	}
+}
+
+func (m *MockParamError) Get(version uint32) (*types.StakingParams, error) {
+	return nil, m.err
 }